reafctored 7b so it also now works with shared IntCode. 7b tests are failing though. Next -> refactor 7a -> fix 7b tests -> clean up D9 -> test D2-5-7-9 are sall still OK

+++ Day05/DiagnosticCode/Program.cs
@@ -12,13 +12,13 @@ namespace Day09
       var input = LoadInput(path);
       
       Executor executor = new Executor(input);
-      executor.InputQueue.Enqueue(1);
+      executor.InputQueue.Enqueue((long)1);
       executor.Execute();
       while (executor.OutputQueue.Count > 0)
         Console.WriteLine("\nPart I: Executor returned: {0}", (long)executor.OutputQueue.Dequeue());
 
       executor.Initialize();
-      executor.InputQueue.Enqueue(5);
+      executor.InputQueue.Enqueue((long)5);
       executor.Execute();
       while (executor.OutputQueue.Count > 0)
         Console.WriteLine("\nPart II: Executor returned: {0}", (long)executor.OutputQueue.Dequeue());
