+++ Day09/SensorBoost/Program.cs
@@ -11,13 +11,13 @@ namespace Day09
       string path = "input.txt";
       Executor executor = new Executor(LoadInput(path));
       executor.Initialize();
-      executor.InputQueue.Enqueue(1);
+      executor.InputQueue.Enqueue((long)1);
       executor.Execute();
       Console.WriteLine("\nPart I :  {0}", (long)executor.OutputQueue.Dequeue());
 
       //reload intcode
       executor = new Executor(LoadInput(path));
-      executor.InputQueue.Enqueue(2);
+      executor.InputQueue.Enqueue((long)2);
       executor.Execute();
       Console.WriteLine("\nPart II :  {0}", (long)executor.OutputQueue.Dequeue());
     }
